# core/vm: change JumpTable []*operation -> []operation