This change changes the definition of `JumpTable` from `[256]*operation` to `[256]operation`. This eliminates one level of pointer indirection so in theory should help with cache locality.