@chfast 2021-12-06
The benchmarks are not really conclusive.

```
name                                                      old time/op  new time/op  delta
EVM/benchmarks/main/blake2b_huff/London/0-8               88.8µs ± 1%  88.1µs ± 1%  -0.77%  (p=0.001 n=10+9)
EVM/benchmarks/main/blake2b_huff/London/1-8               1.92ms ± 0%  1.90ms ± 1%  -1.29%  (p=0.000 n=8+8)
EVM/benchmarks/main/blake2b_huff/London/2-8               3.75ms ± 0%  3.71ms ± 0%  -0.92%  (p=0.000 n=10+8)
EVM/benchmarks/main/blake2b_huff/London/3-8               5.49ms ± 0%  5.43ms ± 0%  -1.13%  (p=0.000 n=8+10)
EVM/benchmarks/main/blake2b_huff/London/4-8               42.7ms ± 0%  42.1ms ± 0%  -1.28%  (p=0.000 n=9+10)
EVM/benchmarks/main/blake2b_shifts/London/0-8             14.2ms ± 0%  14.1ms ± 0%  -1.29%  (p=0.000 n=10+10)
EVM/benchmarks/main/blake2b_shifts/London/1-8             28.4ms ± 1%  27.9ms ± 0%  -1.48%  (p=0.000 n=10+9)
EVM/benchmarks/main/blake2b_shifts/London/2-8             42.5ms ± 1%  41.8ms ± 0%  -1.64%  (p=0.000 n=10+8)
EVM/benchmarks/main/blake2b_shifts/London/3-8              329ms ± 0%   325ms ± 1%  -1.08%  (p=0.001 n=9+9)
EVM/benchmarks/main/sha1_divs/London/0-8                   188µs ± 2%   191µs ± 4%    ~     (p=0.075 n=10+10)
EVM/benchmarks/main/sha1_divs/London/1-8                  3.90ms ± 2%  3.92ms ± 5%    ~     (p=0.356 n=10+9)
EVM/benchmarks/main/sha1_divs/London/2-8                  7.63ms ± 4%  7.71ms ± 0%  +1.02%  (p=0.022 n=10+9)
EVM/benchmarks/main/sha1_divs/London/3-8                  14.9ms ± 2%  14.9ms ± 2%    ~     (p=0.247 n=10+10)
EVM/benchmarks/main/sha1_divs/London/4-8                   180ms ± 2%   181ms ± 2%    ~     (p=0.052 n=10+10)
EVM/benchmarks/main/sha1_shifts/London/0-8                 148µs ± 3%   151µs ± 4%  +2.02%  (p=0.029 n=10+10)
EVM/benchmarks/main/sha1_shifts/London/1-8                3.16ms ± 0%  3.13ms ± 3%    ~     (p=0.515 n=8+10)
EVM/benchmarks/main/sha1_shifts/London/2-8                6.06ms ± 3%  6.07ms ± 3%    ~     (p=0.247 n=10+10)
EVM/benchmarks/main/sha1_shifts/London/3-8                11.9ms ± 3%  11.8ms ± 3%    ~     (p=0.052 n=10+10)
EVM/benchmarks/main/sha1_shifts/London/4-8                 145ms ± 3%   147ms ± 1%    ~     (p=0.669 n=10+7)
EVM/benchmarks/main/weierstrudel/London/0-8                454µs ± 1%   472µs ± 0%  +3.89%  (p=0.000 n=10+10)
EVM/benchmarks/main/weierstrudel/London/1-8                963µs ± 1%   992µs ± 1%  +3.01%  (p=0.000 n=9+10)
EVM/benchmarks/main/weierstrudel/London/2-8               1.53ms ± 0%  1.57ms ± 0%  +2.84%  (p=0.000 n=10+10)
EVM/benchmarks/main/weierstrudel/London/3-8               3.22ms ± 0%  3.30ms ± 1%  +2.52%  (p=0.000 n=9+10)
EVM/benchmarks/main/weierstrudel/London/4-8               4.64ms ± 0%  4.73ms ± 0%  +1.91%  (p=0.000 n=10+10)
EVM/benchmarks/micro/JUMPDEST_n0/London/0-8               2.78ms ± 0%  2.82ms ± 0%  +1.65%  (p=0.000 n=10+9)
EVM/benchmarks/micro/jump_around/London/0-8                123µs ± 0%   125µs ± 0%  +1.76%  (p=0.000 n=9+9)
EVM/benchmarks/micro/loop_with_many_jumpdests/London/0-8  41.9ms ± 1%  43.2ms ± 0%  +3.14%  (p=0.000 n=10+10)
EVM/benchmarks/micro/memory_grow_mload/London/0-8          216µs ± 1%   218µs ± 1%  +1.08%  (p=0.000 n=9+9)
EVM/benchmarks/micro/memory_grow_mload/London/1-8          226µs ± 1%   227µs ± 0%  +0.57%  (p=0.023 n=10+10)
EVM/benchmarks/micro/memory_grow_mload/London/2-8          270µs ± 0%   272µs ± 1%  +0.64%  (p=0.022 n=9+10)
EVM/benchmarks/micro/memory_grow_mload/London/3-8          325µs ± 2%   328µs ± 0%  +0.98%  (p=0.023 n=10+10)
EVM/benchmarks/micro/memory_grow_mstore/London/0-8         443µs ± 1%   444µs ± 0%    ~     (p=0.173 n=10+8)
EVM/benchmarks/micro/memory_grow_mstore/London/1-8         454µs ± 0%   461µs ± 1%  +1.37%  (p=0.000 n=9+10)
EVM/benchmarks/micro/memory_grow_mstore/London/2-8         517µs ± 0%   520µs ± 1%  +0.51%  (p=0.004 n=10+10)
EVM/benchmarks/micro/memory_grow_mstore/London/3-8         555µs ± 1%   568µs ± 0%  +2.31%  (p=0.000 n=10+8)
EVM/benchmarks/micro/signextend/London/0-8                 373µs ± 0%   364µs ± 1%  -2.55%  (p=0.000 n=8+9)
EVM/benchmarks/micro/signextend/London/1-8                 389µs ± 0%   379µs ± 0%  -2.70%  (p=0.000 n=10+10)
[Geo mean]                                                2.47ms       2.48ms       +0.49%
```