@chfast 2021-12-06
The new table layout may not be so great after all. We can observe some increase in L3 cache miss. But this metric is usually chaotic for interpreters.

### old
```
 Performance counter stats for 'go test -o evmbench -run=- -bench=BenchmarkEVM/benchmarks/micro/loop.* -count=10 ./tests':

         12 435,34 msec task-clock                #    1,047 CPUs utilized
            10 039      context-switches          #    0,807 K/sec
               494      cpu-migrations            #    0,040 K/sec
            21 293      page-faults               #    0,002 M/sec
    53 910 096 051      cycles                    #    4,335 GHz                      (30,95%)
   144 346 460 847      instructions              #    2,68  insn per cycle           (38,80%)
    31 340 272 258      branches                  # 2520,259 M/sec                    (39,01%)
        14 282 979      branch-misses             #    0,05% of all branches          (39,14%)
    62 786 788 002      L1-dcache-loads           # 5049,062 M/sec                    (35,91%)
        45 622 717      L1-dcache-load-misses     #    0,07% of all L1-dcache hits    (15,38%)
        11 079 292      LLC-loads                 #    0,891 M/sec                    (15,27%)
           984 777      LLC-load-misses           #    8,89% of all LL-cache hits     (22,87%)
   <not supported>      L1-icache-loads
        16 467 305      L1-icache-load-misses                                         (30,40%)
    62 900 651 237      dTLB-loads                # 5058,219 M/sec                    (27,41%)
         1 146 648      dTLB-load-misses          #    0,00% of all dTLB cache hits   (15,74%)
           702 259      iTLB-loads                #    0,056 M/sec                    (15,13%)
         1 141 537      iTLB-load-misses          #  162,55% of all iTLB cache hits   (22,86%)
   <not supported>      L1-dcache-prefetches
   <not supported>      L1-dcache-prefetch-misses

      11,872302000 seconds time elapsed

      12,321482000 seconds user
       0,216880000 seconds sys
```

### new
```
 Performance counter stats for 'go test -o evmbench -run=- -bench=BenchmarkEVM/benchmarks/micro/loop.* -count=10 ./tests':

         12 352,17 msec task-clock                #    1,047 CPUs utilized
             9 795      context-switches          #    0,793 K/sec
               390      cpu-migrations            #    0,032 K/sec
            21 900      page-faults               #    0,002 M/sec
    53 794 649 990      cycles                    #    4,355 GHz                      (30,77%)
   155 508 357 269      instructions              #    2,89  insn per cycle           (38,48%)
    30 548 701 693      branches                  # 2473,145 M/sec                    (38,47%)
        11 202 157      branch-misses             #    0,04% of all branches          (38,43%)
    57 735 616 042      L1-dcache-loads           # 4674,129 M/sec                    (26,73%)
        39 812 843      L1-dcache-load-misses     #    0,07% of all L1-dcache hits    (17,93%)
        11 571 361      LLC-loads                 #    0,937 M/sec                    (15,10%)
         1 285 150      LLC-load-misses           #   11,11% of all LL-cache hits     (22,77%)
   <not supported>      L1-icache-loads
        21 257 397      L1-icache-load-misses                                         (30,61%)
    57 597 920 547      dTLB-loads                # 4662,981 M/sec                    (29,78%)
         1 571 247      dTLB-load-misses          #    0,00% of all dTLB cache hits   (15,51%)
           657 301      iTLB-loads                #    0,053 M/sec                    (15,44%)
           819 018      iTLB-load-misses          #  124,60% of all iTLB cache hits   (23,13%)
   <not supported>      L1-dcache-prefetches
   <not supported>      L1-dcache-prefetch-misses

      11,799692597 seconds time elapsed

      12,263834000 seconds user
       0,191775000 seconds sys
```